fix: do not use removed View::getLocalFolder()
Actually without replacement. Prior to NC 27 it was implemented on local
storage only, trying to run it would have led into an Exception – if
folders were not filtered out here already. Since the method would
actually only make sense for external folders, removing it seems to be
the reasonable decision, and actually does not remove functionality.

Signed-off-by: Arthur Schiwon <blizzz@arthur-schiwon.de>